backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
openai==1.59.7
google-genai>=1.0.0
numpy==2.2.2
python-dotenv==1.0.1
pydantic==2.10.5
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.2
